refactor(admin): decouple mock data generation from SurveyAnalytics UI
